Moving plDisplayHelper into pl3DPipeline
